clef25/image-retrieval-for-arguments/approaches/bm25_image-caption-baseline/requirements.txt
elasticsearch
python-dotenv